+++ gDRutils/R/merge_SE.R
@@ -44,20 +44,10 @@ merge_SE <- function(SElist,
   metadata <- c(metadata, identifiers)
   
   # 2021.11.01 - param checkDimnames was added to Summarized Experiment in Bioc 3.14, 
-  # to make it compatible with previous solution we set the value to FALSE.
-  se <- if (BiocManager::version() <= "3.13") {
-    SummarizedExperiment(
-      assays = list(
-        Normalized = normalized$BM,
-        Averaged = averaged$BM,
-        Metrics = metrics$BM
-      ),
-      colData = data$condition_md,
-      rowData = data$treatment_md,
-      metadata = metadata
-    )
-  } else {
-    SummarizedExperiment(
+  # to make it compatible with previous solution we re filtering it 'checkDimnames' if not present
+  # TODO: remove once all our envs are with Bioc 3.14
+  p_list <-
+    list(
       assays = list(
         Normalized = normalized$BM,
         Averaged = averaged$BM,
@@ -68,8 +58,10 @@ merge_SE <- function(SElist,
       metadata = metadata,
       checkDimnames = FALSE
     )
-  }
-  se
+  av_pnames <-
+    names(formals(SummarizedExperiment::SummarizedExperiment))
+  f_list <- p_list[intersect(names(p_list), av_pnames)]
+  do.call(SummarizedExperiment, f_list)
 }