+++ gDRutils/tests/testthat/test-identifiers.R
@@ -30,7 +30,7 @@ test_that("reset_env_identifiers works", {
 
 test_that("get_prettified_identifiers works as expected", {
   obs <- get_prettified_identifiers(c("drug_name", "cellline_name"), simplify = FALSE)
-  expect_equal(obs, c("Drug Name", "Cell Line Name"))
+  expect_equal(obs, c("Drug", "Cell Line"))
 
   obs <- get_prettified_identifiers(simplify = TRUE)
   expect_true(is(obs, "list"))
@@ -60,3 +60,4 @@ test_that("get_required_identifiers works as expected", {
   expect_true(length(get_required_identifiers()) > 1L)
   expect_true(all(get_required_identifiers() %in% get_expect_one_identifiers()))
 })
+